+++ src/decorators.ts
@@ -9,6 +9,17 @@
  * rights grant found at http://polymer.github.io/PATENTS.txt
  */
 
+export interface ElementConstructor extends Function {
+  is?: string;
+  properties?: {[prop: string]: PropertyOptions}, observers?: string[];
+  _addDeclarativeEventListener?:
+      (target: string|EventTarget,
+       eventName: string,
+       handler: (ev: Event) => void) => void;
+}
+
+export interface ElementPrototype { constructor: ElementConstructor; }
+
 /**
  * A TypeScript class decorator factory that registers the class as a custom
  * element.
@@ -20,7 +31,7 @@
  * property is not an own-property of the class), an exception is thrown.
  */
 export function customElement(tagname?: string) {
-  return (class_: Function&{is?: string}) => {
+  return (class_: ElementConstructor) => {
     if (tagname) {
       // Only check that tag names match when `is` is our own property. It might
       // be inherited from a superclass, in which case it's ok if they're
@@ -54,25 +65,33 @@ export interface PropertyOptions {
   reflectToAttribute?: boolean;
   readOnly?: boolean;
   computed?: string;
-  observer?: string|((val: any, old: any) => void);
+  observer?: string|((val: {}, old: {}) => void);
 }
 
 function createProperty(
-    proto: any, name: string, options?: PropertyOptions): void {
+    proto: ElementPrototype, name: string, options?: PropertyOptions): void {
   if (!proto.constructor.hasOwnProperty('properties')) {
     Object.defineProperty(proto.constructor, 'properties', {value: {}});
   }
 
   const finalOpts: PropertyOptions = {
-    ...proto.constructor.properties[name] as PropertyOptions | undefined,
+    ...proto.constructor.properties![name] as PropertyOptions | undefined,
     ...options
   };
 
+  interface Reflect {
+    hasMetadata?:
+        (metadataKey: string, proto: object, targetKey: string) => boolean;
+    getMetadata?: (metadataKey: string, proto: object, targetKey: string) =>
+        Object | undefined;
+  }
+
   if (!finalOpts.type) {
-    const reflect = (window as any).Reflect;
-    if (reflect.hasMetadata && reflect.getMetadata &&
+    const reflect = (window as {Reflect?: Reflect}).Reflect;
+    if (reflect && reflect.hasMetadata && reflect.getMetadata &&
         reflect.hasMetadata('design:type', proto, name)) {
-      finalOpts.type = reflect.getMetadata('design:type', proto, name);
+      finalOpts.type = reflect.getMetadata('design:type', proto, name) as
+          PropertyOptions['type'];
     } else {
       console.error(
           `A type could not be found for ${name}. ` +
@@ -80,7 +99,7 @@ function createProperty(
     }
   }
 
-  proto.constructor.properties[name] = finalOpts;
+  proto.constructor.properties![name] = finalOpts;
 }
 
 /**
@@ -90,7 +109,7 @@ function createProperty(
  * This function must be invoked to return a decorator.
  */
 export function property(options?: PropertyOptions) {
-  return (proto: any, propName: string): any => {
+  return (proto: ElementPrototype, propName: string) => {
     createProperty(proto, propName, options);
   }
 }
@@ -102,11 +121,11 @@ export function property(options?: PropertyOptions) {
  * This function must be invoked to return a decorator.
  */
 export function observe(...targets: string[]) {
-  return (proto: any, propName: string): any => {
+  return (proto: ElementPrototype, propName: string) => {
     if (!proto.constructor.hasOwnProperty('observers')) {
       Object.defineProperty(proto.constructor, 'observers', {value: []});
     }
-    proto.constructor.observers.push(`${propName}(${targets.join(',')})`);
+    proto.constructor.observers!.push(`${propName}(${targets.join(',')})`);
   }
 }
 
@@ -120,7 +139,7 @@ export function observe(...targets: string[]) {
  * This function must be invoked to return a decorator.
  */
 export function computed<T = any>(...targets: (keyof T)[]) {
-  return (proto: any,
+  return (proto: ElementPrototype,
           propName: string,
           descriptor: PropertyDescriptor): void => {
     const fnName = `__compute${propName}`;
@@ -169,7 +188,7 @@ export const queryAll = _query(
 function _query(
     queryFn: (target: NodeSelector, selector: string) =>
         Element | NodeList | null) {
-  return (selector: string) => (proto: any, propName: string): any => {
+  return (selector: string) => (proto: ElementPrototype, propName: string) => {
     Object.defineProperty(proto, propName, {
       get(this: HTMLElement) {
         return queryFn(this.shadowRoot!, selector);
@@ -194,7 +213,13 @@ function _query(
  * @param target A single element by id or EventTarget to target
  */
 export const listen = (eventName: string, target: string|EventTarget) =>
-    (proto: any, methodName: string) => {
+    (proto: ElementPrototype, methodName: string) => {
+      if (!proto.constructor._addDeclarativeEventListener) {
+        console.error(
+            `Cannot add listener for ${eventName} because ` +
+            `DeclarativeEventListeners mixin was not applied to element.`);
+        return;
+      }
       proto.constructor._addDeclarativeEventListener(
-          target, eventName, proto[methodName]);
+          target, eventName, (proto as any)[methodName]);
     };