+++ CHANGELOG.md
@@ -1,9 +1,9 @@
 # Change Log
 
-<!-- ## [Unreleased] -->
+## [Unreleased]
 - An exception is now thrown if both the `tagname` parameter to `@customElement` is provided, and the static `is` class property is set (except in the case that the `is` property is not an own-property of the class). Previously the `is` property would always silently win over the `tagname` when both were set.
+- An exception is now thrown if the `@listen` decorator is applied to an element that does not have the `DeclarativeEventListeners` mixin applied.
 - Decorators now have more constrained type signatures.
-- `@listen` decorator will now warn if the `DeclarativeEventListeners` mixin was not applied.
 
 ## [1.1.1] - 2018-02-16
 - Remove npm dependency on `reflect-metadata` package.